BaiTapLonLTHD/alumni/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-ckeditor==6.7.2
django-debug-toolbar==5.0.1
django-js-asset==3.0.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
